requirements.txt
discord.py>=2.3.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
Pillow>=10.0.0
edge_tts>=7.2.7